+++ ktuvitDownloader/__main__.py
@@ -2,12 +2,17 @@
 # -*- coding: utf-8 -*-
 import ConfigParser
 import base64
+import json
 import logging
 import os
 from getpass import getpass
 from logging import handlers
 
 import sys
+
+from datetime import datetime
+import yaml
+from appdirs import AppDirs
 from guessit import guessit
 
 from ktuvitDownloader.CustomExceptions import CantFindSubtitleException, WrongLoginException
@@ -17,7 +22,12 @@ from ktuvitDownloader.const import *
 from ktuvitDownloader.files import get_paths_files, move_finshed
 from ktuvitDownloader.options import args_parse, parse_log
 
+
 config = ConfigParser.RawConfigParser()
+app_dir = AppDirs("ktuvitdownloader")
+config_file = os.path.join(app_dir.user_data_dir, CONFIG_FILE)
+log_file = os.path.join(app_dir.user_data_dir, LOG_FILE)
+cache_file = os.path.join(app_dir.user_data_dir, CACHING_FILE)
 
 
 def main():
@@ -25,6 +35,7 @@ def main():
     main function
     """
     options = args_parse.parse_args()
+    check_user_data_dir()
     global logger
     logger = logging.getLogger()
     logger.addHandler(get_handler(not options.specific))
@@ -33,7 +44,7 @@ def main():
     base_dir = ""
     dest_dir = ""
 
-    if options.version:
+    if options.show_version:
         print_version()
         exit(0)
 
@@ -45,11 +56,15 @@ def main():
         print_log(isShort=False)
         exit(0)
 
-    if options.configuration:
+    if options.show_configuration:
         print_configuration()
         exit(0)
 
-    if not os.path.isfile(CONFIG_FILE):
+    if options.show_cache:
+        print_cache()
+        exit(0)
+
+    if not os.path.isfile(config_file):
         options.reset = True
 
     if options.reset:
@@ -69,7 +84,7 @@ def main():
                 file_name = base_dir.split("\\")[-1]
                 base_dir = os.path.dirname(os.path.realpath(base_dir))
                 dest_dir = os.path.abspath(os.path.join(base_dir, os.pardir))
-                config.read(CONFIG_FILE)
+                config.read(config_file)
                 username = config.get("Login", "username")
                 password = base64.b64decode(config.get("Login", "password"))
                 downloader(base_dir, dest_dir, {os.path.join(base_dir, file_name): guessit(file_name)}, username,
@@ -87,7 +102,7 @@ def main():
         except ConfigParser.DuplicateSectionError:
             pass
         config.set("Directories", "base_dir", base_dir)
-        with open(CONFIG_FILE, "wb") as f:
+        with open(config_file, "wb") as f:
             config.write(f)
 
     if options.dest_path:
@@ -100,10 +115,10 @@ def main():
             config.set("Directories", "dest_dir", dest_dir)
         except ConfigParser.DuplicateSectionError:
             pass
-        with open(CONFIG_FILE, "wb") as f:
+        with open(config_file, "wb") as f:
             config.write(f)
 
-    config.read(CONFIG_FILE)
+    config.read(config_file)
     username = config.get("Login", "username")
     password = base64.b64decode(config.get("Login", "password"))
     if not options.specific:
@@ -115,35 +130,59 @@ def main():
     downloader(base_dir, dest_dir, files, username, password, options.specific)
 
 
-def print_configuration():
-    if not os.path.isfile(CONFIG_FILE):
+def check_user_data_dir():
+    app_dir_list = app_dir.user_data_dir.split("\\")
+    for i in range(1, len(app_dir_list)):
+        if not os.path.exists("\\".join(app_dir_list[:i + 1])):
+            os.mkdir("\\".join(app_dir_list[:i + 1]))
+
+
+def print_cache():
+    def json_serial(obj):
+        """JSON serializer for objects not serializable by default json code"""
+
+        if isinstance(obj, datetime):
+            serial = obj.isoformat()
+            return serial
+        raise TypeError("Type not serializable")
+
+    if not os.path.isfile(cache_file):
         print "Nothing to show"
     else:
-        with open(CONFIG_FILE) as f:
+        with open(cache_file) as f:
+            text = yaml.load(f)
+        if not text:
+            print "Nothing to show"
+        else:
+            print json.dumps(text, default=json_serial, indent=4, ensure_ascii=False)
+
+
+def print_configuration():
+    if not os.path.isfile(config_file):
+        print "Nothing to show"
+    else:
+        with open(config_file) as f:
             text = f.read()
-            if not text:
-                print "Nothing to show"
-            else:
-                print text
-                config.read(CONFIG_FILE)
-                c_password = base64.b64decode(config.get("Login", "password"))
-                print "Decoded password:", c_password
+        if not text:
+            print "Nothing to show"
+        else:
+            print text
 
 
 def print_log(isShort):
-    if not os.path.isfile(LOG_FILE):
+    if not os.path.isfile(log_file):
         print "Nothing to show"
     else:
-        with open(LOG_FILE) as f:
+        with open(log_file) as f:
             if isShort:
                 text = f.readlines()
                 text = parse_log(text)
             else:
                 text = f.read()
-            if not text:
-                print "Nothing to show"
-            else:
-                print text
+        if not text:
+            print "Nothing to show"
+        else:
+            print text
 
 
 def print_version():
@@ -158,7 +197,7 @@ def print_version():
 def get_handler(isFile):
     formatter = logging.Formatter(fmt="%(asctime)s\t%(levelname)s: %(message)s")
     if isFile:
-        handler = handlers.RotatingFileHandler(filename=LOG_FILE, maxBytes=MB)
+        handler = handlers.RotatingFileHandler(filename=log_file, maxBytes=MB)
     else:
         handler = logging.StreamHandler(sys.stdout)
     handler.setFormatter(formatter)
@@ -167,7 +206,7 @@ def get_handler(isFile):
 
 def downloader(base_dir, dest_dir, files, username, password, specific=False):
     downloaded = []
-    con = Connection(username, password)
+    con = Connection(username, password, app_dir)
     try:
         con.login()
     except WrongLoginException as e: