+++ ktuvitDownloader/const.py
@@ -14,7 +14,8 @@ LOGIN_BLOCKED = "המשתמש שברשותך אינו פעיל"
 SUB_EXT = [".srt", ".sub"]
 VIDEO_EXT = [".webm", ".mkv", ".flv", ".avi", ".mov", ".wmv", ".rm", ".rmvb", ".mp4", ".m4p", ".m4v", ".mpg", ".mpeg",
              "mp2", ".mpe", ".mpv", ".m2v", ".m4v"]
-TORRENTS_GROUPS = ["[ettv]", "[eztv]"]
+TORRENTS_GROUPS = ["[ettv]", "[eztv]", "[rartv]", "[P2PDL]", "[TJET]", "[EtHD]", "[PRiME]", "[VTV]", "[DDR]",
+                   "[DSRG]", "[UTR]", "[GloDLS]"]
 
 CONFIG_FILE = ".ktuvitConfig.cfg"
 LOG_FILE = ".ktuvitLogger.log"
@@ -22,3 +23,5 @@ CACHING_FILE = ".ktuvitCache.yml"
 
 MB = 1024 * 1024
 KB = 1024
+
+CACHE_DAYS = 8